requirements.txt
lightning>=2.0.0
torchmetrics>=0.11.4
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
tensorboard
librosa
Cython
numpy
einops
inflect
Unidecode
scipy
torchaudio
matplotlib
pandas
conformer==0.3.2
diffusers 
s3prl==0.4.10
transformers==4.28.0
tiktoken